seperation/ConvTasNet/requirements.txt
tensorflow
tensorboard
soundfile
librosa
tqdm
pyyaml